🐞: Fix meet.google.com Webcam
There is a "hack" to allow google login which replaces the Electron/Chromium based
User-Agent with a fake FireFox user-agent.

This is causing trouble with the latest release of Google Meet.
The current fix replaces the logic for setting user-agent in this workaround and
adds a specific case to bypass meet.google.com